chore: Add pre-commit hooks to run CI locally
This `pre-commit-config` tries to simulate locally all the checks that
are run in the already configured CI, avoiding pushing code all the time
to the remote server to only then make sure if the code is working or
not. This change, however, is not meant to replace the checks done at
CI, its sole purpose is to help developers in their workflow process.
The final word when it comes to approving code changes or not is only
through the CI.